# Conflicts:
#	src/main/resources/data.sql

+++ src/main/java/tech3/binitright/repository/AdminRepository.java
@@ -6,6 +6,7 @@ import tech3.binitright.model.Admin;
 import java.util.Collection;
 import java.util.List;
 
-public interface AdminRepository extends JpaRepository<Admin, Integer> {
+public interface AdminRepository extends JpaRepository<Admin, Long> {
     List<Admin> findByUsername(String username);
+
 }
